Hotfix Release 1.0.1: ABS dependency removed
In Release 1.0.0, the BaseFragment class previously extended
SherlockFragment, and by that dragged in the whole ActionBarSherlock
Library. As this can have undesirable side-effects, this hotfix
was release to remove this problem.
BaseFragment extends the support-v4 package's Fragment-class directly
now, and thereby remove the necessity for dragging in the ABS library.
If you need the functionality of the BaseFragment but with the features
the SherlockFragment provides as well, take the source-code from
BaseFragment, copy the class into your own project, and change the
line "implements Fragment" to "implements SherlockFragment". Through the
benefits of being an open source library, this should not really pose a
problem, just a minor inconvenience. :)